+++ database/migrations/2023_09_11_075854_create_component_product_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId("component_id")->constrained("components")->onDelete("cascade");
             $table->foreignId("product_id")->constrained("products")->onDelete("cascade");
-            $table->integer("quantity");
+            $table->double("quantity");
             $table->timestamps();
         });
     }
